[Refactor] Divide files
