+++ zero2prod/src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod health_check;
+mod subscriptions;
+
+pub use health_check::*;
+pub use subscriptions::*;
