+++ zero2prod/tests/health_check.rs
@@ -1,6 +1,6 @@
 use std::net::TcpListener;
+use zero2prod::app;
 use zero2prod::config;
-use zero2prod::run;
 
 use sqlx::{Connection, PgConnection};
 
@@ -8,7 +8,7 @@ fn spawn_app() -> String {
     let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
     // We retrieve the port assigned to us by the OS
     let port = listener.local_addr().unwrap().port();
-    let server = run(listener).expect("Failed to bind address");
+    let server = app::run(listener).expect("Failed to bind address");
     let _ = tokio::spawn(server);
     // We return the application address to the caller!
     format!("http://127.0.0.1:{}", port)
